+++ .env
@@ -1,6 +1,6 @@
 APP_NAME=Laravel
 APP_ENV=local
-APP_KEY=base64:1UPWQrRU6Mg6nnFA3s9LHTFysLjjVXql3yTZXcisFao=
+APP_KEY=base64:4AUsc7081ASb7t5w3nVsWkcDkLmqzpLWTRvx91p3Xd4=
 APP_DEBUG=true
 APP_URL=http://localhost:8080
