+++ app/Http/Controllers/AlunoLoginController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class AlunoLoginController extends Controller
+{
+   
+    public function index() {
+        return view('events.areadoaluno');
+    }
+
+   
+}
+
