+++ app/Http/Controllers/PaginaInicialController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class PaginaInicialController extends Controller
+{
+    public function index(){
+        return view('events.welcome');
+    }
+
+
+
+}
