+++ app/Http/Controllers/ProfessorLoginController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class ProfessorLoginController extends Controller
+{
+   
+    public function index() {
+        return view('events.areadoprofessor');
+    }
+
+   
+}
