+++ app/Http/Controllers/SecretariaLoginController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class SecretariaLoginController extends Controller
+{
+   
+    public function index() {
+        return view('events.areadasecretaria');
+    }
+
+   
+}
