--- app/Http/Controllers/alunos.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-use App\Model\Aluno;
-
-class alunos extends Controller
-{
-    public function index(){
-
-        $alunos = Aluno::all();
-
-        return view('welcome',['alunos' => $alunos])
-    }
-}
